+++ app/core/tests/functional/model_tickets_meta/test_functional_model_tickets_model_meta_models.py
@@ -153,8 +153,8 @@ class ModelTicketMetaModelTestCases(
 
 
                 content_type = model_contenttype.objects.filter(
-                    app_label = ticket_model._meta.app_label,
-                    model = ticket_model._meta.model_name
+                    app_label = model._meta.app_label,
+                    model = model._meta.model_name
                 ).first()
 
                 model_kwargs['content_type'] = content_type