+++ app/core/tests/unit/ticket_base/test_unit_ticket_base_viewset.py
@@ -58,6 +58,7 @@ class ViewsetTestCases(
             'filterset_fields': {
                 'value': [
                     'is_deleted',
+                    'milestone',
                     'organization',
                     'project',
                 ]