+++ app/core/tests/unit/ticket_dependency/test_unit_ticket_dependency_model.py
@@ -45,12 +45,10 @@ class TicketDependencyModelTestCases(
                     "dataset": {
                         "columns": [
                             [
-                                'id',
-                                'title',
-                                'status_badge',
-                                'opened_by',
-                                'organization',
-                                'created'
+                                'how_related_icon'
+                            ],
+                            [
+                                'display_name'
                             ]
                         ]
                     },