+++ includes/etc/itsm/settings.py
@@ -3,6 +3,24 @@
 
 # ADDITIONAL_APPS: list = []    # Any additional Django apps to install
 
+CSRF_COOKIE_PATH = '/'                # If hosting on a different path, this must be updated to match.
+
+#
+# Note: If using SSO, the cookie age here must be set to a few seconds less than the SSO providers
+#       cookie / token (refresh) expiry. This allows for Centurion ERP to force reauthentication.
+#
+SESSION_COOKIE_AGE = ( 3600 * 4 )     # 4 hours. Age the session cookie should live for in seconds.
+SESSION_COOKIE_DOMAIN = None          # This must be set to the domain that Centurion ERP is hosted on.
+SESSION_COOKIE_PATH = '/'             # If hosting on a different path, this must be updated to match.
+
+#
+# Note: This cookie age should be slightly longer than the session cookie age. This allows for the
+#       session cookie to expire and ensures that the CSRF cookie is always valid for the session
+#       duration.
+#
+CSRF_COOKIE_AGE = ( SESSION_COOKIE_AGE + 2 )     # Age the CSRF cookie should live for in seconds.
+
+
 #
 # If metrics enabled, see https://nofusscomputing.com/projects/centurion_erp/administration/monitoring/#django-exporter-setup)
 # to configure the database metrics.
@@ -48,9 +66,13 @@ LOG_FILES = {    # Location where log files will be created
 }
 
 
+SECRET_KEY = None    # You must generate this
+
 SECURE_PROXY_SSL_HEADER = ("HTTP_X_FORWARDED_PROTO", "https")
 
 SECURE_SSL_REDIRECT = True
+
+SITE_URL = 'http://127.0.0.1'    # Base URL the site is hosted on.
 
 # TRACE_LOGGING = True                                                                # Enable Trace Logging.
 # CENTURION_LOGGING['loggers']['centurion.trace']['level'] = CenturionLogger.TRACE    # Set Trace Logging level. normally not required