+++ Release-Notes.md
@@ -1,3 +1,14 @@
+## Version v1.33.0
+
+- remove feature flag `2025-00006`.
+
+- place model Request Ticket behind feature flag `2026-00012` until model ready for production use.
+
+- Support added whe updating role permissions to use format `<app>.<permission>_<model>`
+
+- This release is the first release that provides the routes to [Centurion UI](https://nofusscomputing.com/projects/centurion_user_interface/describing/#ui-routes). As such [Centurion UI v0.14.0](https://github.com/nofusscomputing/centurion_ui/pull/243) will be required for Centurion ERP to work.
+
+
 ## Version v1.32.0
 
 - New Models